+++ client/src/components/views/LoginPage/LoginPage.css
@@ -43,3 +43,7 @@
     top: 5%;
     margin-top: 10px;
 }
+
+#auth-btn{
+    margin-top: 100px !important;
+}
