+++ client/src/components/views/LoginPage/LoginPage.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import { withRouter } from 'react-router-dom';
 import {useDispatch} from 'react-redux';
 import {loginUser} from '../../../_actions/user_action'
@@ -6,59 +6,34 @@ import '../views.css';
 import './LoginPage.css';
 import 'bootstrap/dist/css/bootstrap.css';
 
-import axios from 'axios'
 
+import LoginForm from './LoginForm'
 
 function LoginPage(props) {
-
-  const dispatch = useDispatch();
-
-  const [Email, setEmail] = useState("");
-  const [Password, setPassword] = useState("");
   
-  const onEmailHandler = (event) => {
-    setEmail(event.currentTarget.value);
-  }
-  
-  const onPasswordHandler = (event) => {
-    setPassword(event.currentTarget.value);
+  const [auth, setAuth] = useState(null)
+
+  useEffect(()=>{
+    setAuth(null)
+  }, [])
+  const onStudentClick = ()=>{
+    setAuth('student');
   }
 
-  const onSubmitHandler = (event) =>{
-    event.preventDefault();
-
-    let userBody = {
-      email: Email,
-      password: Password
-    }
-    // server 연결 시 userBody fetch. 현재는 state 확인 용으로 console log
-    console.log(userBody);
-  
-    axios.post('/app/login/students', userBody)
-    .then(response=>{
-      console.log(response)
-    })
-    .catch(err=>console.log(err))
+  const onProfessorClick = () =>{
+    setAuth('professor');
   }
 
   return (
     <div id="wrap">
       <div id="login-content">
         <h1 id="ment">Sign in</h1>
-        <div id="form-box">
-          <form id="login-form" onSubmit={onSubmitHandler}>
-            <div id="id-group" className="form-group">
-              <label for="uid">이메일</label>
-              <input type="email" className="form-control" id="uid" name="uid" value={Email} onChange={onEmailHandler} required/>
-            </div>
-            <div className="form-group">
-              <label for="pwd">Password</label>
-              <input id="pwd" className="form-control" name="pwd" type="password" value={Password} onChange={onPasswordHandler} required/>
-            </div>
-            <button type="submit" id="login-btn" className="btn btn-primary">Sign in</button>
-          </form>
-        </div>
-      </div>
+      {!auth && <>
+      <button id="auth-btn" onClick={onStudentClick}>학생 로그인</button>
+      <button id="auth-btn" onClick={onProfessorClick}>교수 로그인</button>
+      </>}
+      {auth && <LoginForm auth={auth}/>}
+    </div>
     </div>
   );
 }
